Extract sub-authorities to separate classes This helps pave the way for adding local authorities that are not backed by a yaml file.
This is not backward compatible so it should be 0.5.0
