+++ lib/qa/authorities/oclcts.rb
@@ -2,45 +2,24 @@ require 'open-uri'
 require 'nokogiri'
 
 module Qa::Authorities
-  class Oclcts < AuthorityWithSubAuthority
-    include WebServiceBase
+  module Oclcts
+    require 'qa/authorities/oclcts/generic_oclc_authority'
+    extend AuthorityWithSubAuthority
+
+    def self.factory(sub_authority)
+      validate_sub_authority!(sub_authority)
+      GenericOclcAuthority.new(sub_authority)
+    end
 
     SRU_SERVER_CONFIG = YAML.load_file(Rails.root.join("config", "oclcts-authorities.yml"))
 
-    def sub_authorities
+    def self.sub_authorities
       SRU_SERVER_CONFIG["authorities"].map { |sub_authority| sub_authority[0] }
     end
 
-    def search q
-      get_raw_response("prefix-query", q)
-      r = Array.new
-      raw_response.xpath('sru:searchRetrieveResponse/sru:records/sru:record/sru:recordData', 'sru' => 'http://www.loc.gov/zing/srw/').each do |record|
-        r.append({"id" => record.xpath('Zthes/term/termId').first.content, "label" => record.xpath('Zthes/term/termName').first.content})
-      end
-      r
+    def self.url_pattern(query_type)
+      SRU_SERVER_CONFIG["url-pattern"][query_type]
     end
 
-    def find id
-      get_raw_response("id-lookup", id)
-      parse_full_record(id)
-    end
-
-    private
-
-    def parse_full_record id
-      a = {}
-      zthes_record = raw_response.xpath("sru:searchRetrieveResponse/sru:records/sru:record/sru:recordData/Zthes/term[termId='#{id}']", 'sru' => 'http://www.loc.gov/zing/srw/')
-      zthes_record.children.each do |child|
-        if (child.is_a? Nokogiri::XML::Element) && (!child.children.nil?) && (child.children.size == 1) && (child.children.first.is_a? Nokogiri::XML::Text)
-          a[child.name] = child.children.first.to_s
-        end
-      end
-      a
-    end
-
-    def get_raw_response query_type, id
-      query_url = SRU_SERVER_CONFIG["url-pattern"][query_type].gsub("{query}", id).gsub("{id}", id).gsub("{authority-id}", sub_authority)
-      @raw_response = Nokogiri::XML(open(query_url))
-    end
   end
 end
