+++ spec/lib/authorities_getty_spec.rb
@@ -3,24 +3,30 @@ require 'spec_helper'
 describe Qa::Authorities::Getty do
 
   describe "#new" do
+    it "should raise an exception" do
+      expect { described_class.new }.to raise_error
+    end
+  end
+
+  describe "#factory" do
     context "without a sub-authority" do
       it "should raise an exception" do
-        expect { described_class.new }.to raise_error
+        expect { described_class.factory }.to raise_error
       end
     end
     context "with an invalid sub-authority" do
       it "should raise an exception" do
-        expect { described_class.new("foo") }.to raise_error
+        expect { described_class.factory("foo") }.to raise_error
       end
     end
     context "with a valid sub-authority" do
       it "should create the authority" do
-        expect(described_class.new("aat")).to be_kind_of described_class
+        expect(described_class.factory("AAT")).to be_kind_of Qa::Authorities::Getty::AAT
       end
     end
   end
 
-  let(:authority) { described_class.new("aat") }
+  let(:authority) { described_class.factory("AAT") }
   describe "#build_query_url" do
     subject { authority.build_query_url("foo") }
     it { is_expected.to  match /^http:\/\/vocab\.getty\.edu\// }
